Add comparison of hands
Implements functionality for calculating the rank of a hand, and comparing two hands, returning the hand which has the best hand rank. If ranks are equal, we check which hand has the highest card. If that's also equal, we return null indicating that no hand is better than the other.